changed view model to adhere to abstract view model class, add setters and getters for state class
